+++ MainGame/objects/Enemy.hpp
@@ -31,7 +31,7 @@ public:
     auto getDisplayPosition() const
     {
         auto vec = getPosition();
-        return sf::Vector2f(std::floor(vec.x), std::floor(vec.y));
+        return sf::Vector2f((float)std::floor(vec.x), (float)std::floor(vec.y));
     }
 
     static constexpr cpCollisionType CollisionType = 'enmy';
